requirements.txt
beautifulsoup4==4.9.0
python-telegram-bot==12.5.1
requests==2.23.0
m3u8==0.5.4
ffmpeg-python==0.2.0
tgcrypto==1.2.0
aiofiles==0.5.0
git+git://github.com/kfur/Telethon@master
git+git://github.com/kfur/python-zipstream@async
logaugment==0.1.3
urlextract==0.14.0
aiohttp==3.6.2
aiogram==2.8
Pillow==7.2.0
cchardet==2.1.6
pycares==3.1.1
aiodns==2.0.0
psycopg2-binary==2.8.5
selenium==3.141.0
youtube-dl==2021.02.04.1
